CCP-135 RDF format for API (#120)
* Initial change

* Changes for generating rdf file

* RDF Changes

* Deleting unwanted files

* Initial change

* Changes for generating rdf file

* Conflicts fix

* review comment changes for rdf

* Initial change

* Changes for generating rdf file

* RDF Changes

* Deleting unwanted files

* Merge Conflicts fix

* Changes for merge conflict

* Changes for lucene search

* Changes for review comments

* [CCP-135] Compilation error changes

* [CCP-135] Removed class path changes

* [CCP-135] Removed class path

* [CCP-135] Added bad uri exception

Bad URI exception is added to make sure when a bad uri
occurs, the response is not crashed. Application returns the
results of the correct uris.

* [CCP-135] Fixed bad uri exception

This exception was occurring because the create resource had an empty
parameter. Create resource can have null as parameter but not empty
Second round review comments fix: Renamed the method name.
Moved setting NsPrefix outside the loop.

* [CCP-135] Review comments changes