[CCP-168] Form will not be submitted when enter is pressed (#131)
* [CCP-168] Form will not be submitted when enter is pressed

Following changes are made to concept add screen.
1) When user does not enter any value for Term field
and tries to click on search, we do not hit the database,
instead we display error message.
2) In case of any error during ajax call we display
processing error to the user.
3) When user tries to search a word that is not in VIAF an
unknown alert box was appearing on the screen. Now handled
that condition by clearing the data table and displaying proper error
message.

* [CCP-168] Made sure when enter key is pressed on Term field
only search happens. When enter key is pressed on other
fields the form will be submitted.

* [CCP-168] In case of error message the processing
image is removed. Else the ajax processor image
will be displayed even after validation.

* [CCP-168] Formatting changes