[CCP-176] Changes for pagination through rest interface
Pagination is carried out through lucene utility. If user searches with pagenumber
as 1, the system fetches the first 10 records and also provides the total number of
search result for that particular search. User can specify the page number through
"page" request parameter. From second page we fetch the corresponding records
and display.

If user tries to fetch the record from 5th page and total number of records is 40 the
system displays an empty result meaning there is no result available for 5th page.