[CCP-193] Rest call search fix
There were two issues in rest ful call concept search
1) The type_id has been included in the query even when
type_id is null. When type_id is null, the old code used a
String.valueOf so "null" string was included in query string.
This is now removed checking null for types field like all other
fields.

2) The default search operator was OR for concept search rest
call, where as the search for UI is AND. Made sure both the search
use AND for default operator. If operator is supplied by user, then
that operator will be overridden.

3) There was one more issue with operator. The operators supplied
to lucene query string should always be in upper case. So made changes
accordingly.