[CCP-193] Fixed ordering of results in rest calls
There was discrepancy between search results in UI and rest
call. This is because sorting was performed in UI which is now
disabled on page load. And in rest call the concept entry is loaded
from arraylist to hashMap, Hashmap does not retain the ordering.
So hash map is replaced with LinkedHashMap.

Formatting changes on constant file