[CCP-186] Changes for handling word net ids
Wordnet ids cannot be deleted. So when word net ids are
merged, we create a concept wrapper and delete the wrapper.
If user wants to merge only word net concepts, user is provided
with the option to create a new id. If the merging concepts
contain CCP concept, then that id is provided in the list.
This is because we cannot update the word net concepts.