story/CCP-187 (#149)
* [CCP-187] Changes for deletion updates

When a deleted concept is searched by id, the
deleted concept will not be retrieved. When a
deleted concept is searched after reindexing, the
deleting concept will not be shown as well.

* [CCP-187] Reindexing fix for deleted concepts.

Word net ids which are linked to wrappers are not
required to be indexed. This is removed by getting all
the word net ids related to concept wrapper and not
indexing.

* [CCP-187] Review changes

Removed the deletion condition check from ConceptManager
and moved it to Controller. Removed logger statement from
lucene utility.

* [CCP-187] Deleted concepts will be retrieved in lookup by id

Deleted concepts will have the deleted flag set to true in json and xml