[CCP-176] Root element is added correctly. (#153)
* [CCP-176] Root element is added correctly.

Previously pagination was added after root element. Just updated
the root element.

* [CCP-176] Review comments changes.

Added a validator to check if the returned xml is valid or not in test cases.
Made sure correct xml is returned when pagination object is not passed to
the jsonconceptmessage or xmlconceptmessage.
Removed the getConceptEntries method and all the json and xml messages
will be returned by getConceptEntriesAndPagination method. If the pagination
object is passed on null then the method acts as getConceptEntries.