Story/CCP-186 (#148)
* [CCP-186] Initial commit

* [CCP-186] Initial commit.

Need to decide on generating the checkboxes.

* [CCP-186] Changes made in UI for merging

1) UI changes for merging concepts. User selects the concepts
to be merged and clicks on merge.
2) Once button is clicked, on server side we fetch the values for
all the selected concept ids and merge. Merge functionality is done
by removing the duplicates. Removing duplicates and tuning activites
still pending.
3) After tuning the data, the merged entry is provided in the UI for verification
by the user. Some work pending in this screen as well.

* [CCP-186] Added backend logic for merging

Backend logic is included in service class. Need
to test the changes. UI tweaking is required to complete
the story.

* [CCP-186] Added word net ids.

If word net ids are merged, the word net ids are not deleted. because
the word net ids are retrieved from wordnetmanager and the delete
method in concept manager is not supporting the deletion of word net
ids. So instead of deleting, these word net ids are added to the word net
field of the concept mergebean

* [CCP-186] UI Changes.

Made changes for not displaying merge entry at first.
When prepare to merge is clicked, edit and delete are
removed and merge checkbox is included. Added error
messages as well for validation.

* [CCP-186] Issue fix for lucene utility

* [CCP-186] Formatting changes

* [CCP-186] Changes for review comments.

Used lamdas for creating string from set<String>
Changed the variables names. Used plural form for collection
variables.

* [CCP-186] Changed java version to 1.8

* [CCP-186] Changes for handling word net ids

Wordnet ids cannot be deleted. So when word net ids are
merged, we create a concept wrapper and delete the wrapper.
If user wants to merge only word net concepts, user is provided
with the option to create a new id. If the merging concepts
contain CCP concept, then that id is provided in the list.
This is because we cannot update the word net concepts.

* [CCP-186] Review comments changes

Formatting changes and review changes on merging
code. Changed for request mapping.

* [CCP-186] Review comments changes

Instead of having a single object. Input values are
loaded onto model object and output/submitted values
are stored in the ConceptsMergeBean object.

* [CCP-186] Review comments fix

Moved pos values to be fetched from the message file.
Used form:radiobuttons instead of iterating the radio buttons
Used filter in streams. Formatted Lucenefieldname
and searchfiledname

* [CCP-186] Review comments changes

Changed the variable name from p to id.

* [CCP-186] Addressed review comments.

For loading the POS drop down spring message is used and
loaded from the properties file in resource folder. When there
are no wrapper concepts,  a message is shown to the user stating
a new id will be created for the merged concepts.

* [CCP-186] Review comments fix

Used streams in most of the code. Changed the request mapping.
Removed unwanted filter conditions.

* [CCP-186] Added validation checks for concept merge

* [CCP-186] Changed the java version in maven.

Changed the method name as per comments

* [CCP-186] Review comments changes

Changes for displaying word, pos and type values.
Warning message will be displayed if there is  a change in pos
or concept types between merging concepts. Need to fix description
review comments.

* [CCP-186] Changes for review comments.

Description and word have been displayed as per review comments
Warning message is displayed when there is a difference in pos
or concept type.
Rearranged the buttons in UI.
Description is displayed as paragraphs.

* [CCP-186] Fix for something went wrong.

Wordnet ids were not deleted correctly. Fixed by deleting
wordnet ids one by one in lucene.

* [CCP-186] Changes for description field.

Instead of adding the html tags in the java end. The descriptions are
loaded on to a list and iterated in jsp with paragraph tags.

* [CCP-186] Changes for adding the merged ids to new concept.

Previously word net ids were added to word net field. But from now
on all ids will be added to merged ids and will not be added to
alternative ids field as well.

* [CCP-186] Included error class for displaying errors.

* [CCP-186] Changes for review comments

Error message is loaded from locale file.
Buttons are provided with padding to prevent merging with
tables.
Made sure error messages are correctly added for pos and types

* [CCP-186] isMultiple changes for concept merge.

Merge ids added as a multiple field with tokenized to false.

* [CCP-186] ID changes for concept merge

Alternative ids and merged ids has been set based on the excel
sheet in drop box. Wordnet id will always be blank when concepts
are getting merged. word net ids will be pushed to alternative ids
in case of merge.

* [CCP-186] Review comments changes.

1) Wordnet id has already been set in word net concepts. So removed the
setter for wrapper ids
2) Made sure validator is working fine. Previously the values for drop down
has not been loaded and redirect was used. Did not use redirect because on
load performs some operations which is not required while validating.

* [CCP-186] Review comments changes

* [CCP-186] Review changes.

Removed the map and added the entire concept entries to be iterated
in jsp.

* [CCP-186] Review comments changes

Changed the validation to validate white space as well.
Changed the button names.
Added validation to check if merged ids contain at least two ids.
Merged concept details are displayed as per review comments.

* [CCP-186] Review comments changes

Removed conceptIds field and conceptwordnetids field from bean
and added to model object.

* [CCP-186] Changes for including the null check.

Merged ids will be validated for null value. If the value is not null,
then we check if the merged ids are more than 2. If not we throw
validation error,

* [CCP-186] Removed _ from words and trimmed description

* [CCP-186] Review comments changes

Removed the space in url while merging. The space causes double encoding
while forwarding from login. Added the _ replace in words in word net manager
instead of adding the replace in addConceptEntry(). Removed period from button.