[CCP-202] Review comments changes
Added new class for separating servlet context and rest context. But was not able
to create two different classes, because restful calls also needed lucene index which
is available in servlet context.

Added database files to the workspace and read from workspace.

Added test case for the first restful call and validated the output from server.