[CCP-211] Removed the backslash "/" in concept entry
Back slash has been removed in equalsTo and similarTo in concept entry. This has been
removed in setter because lucene gets the values from the object using reflection and not
the getters so changes are made while loading the values. Also from now on any insertion of
equalsTo and simialrTo will not be having "/".

Also any search with "/" at the end needs to be removed. This will be done using spring conversion
service.