Bug/CCP-211 (#167)
* [CCP-211] Removed the backslash "/" in concept entry

Back slash has been removed in equalsTo and similarTo in concept entry. This has been
removed in setter because lucene gets the values from the object using reflection and not
the getters so changes are made while loading the values. Also from now on any insertion of
equalsTo and simialrTo will not be having "/".

Also any search with "/" at the end needs to be removed. This will be done using spring conversion
service.

* [CCP-211] Changes for concept search of equals and similar to

Before searching for equals and similar to, we can check if equals
and similar to contains a back slash in the end of the url. If so we
remove the backslash and search on lucene. Also while inserting the
values on to lucene we check for the backslash.

* [CCP-211] Formatting changes

* [CCP-211] Review comments changes

Handled multiple equalsTo and similarTo string values.
used PropertyDescriptor for getting and setting values.
Included exceptions in unit test cases.

* [CCP-211] Review comments changes.

Moved operators, similarTo and equalsTo to bean class. pre processing of
typeUri and numberOfPages in concept search class. This is because
typeUri and numberOfpAges needs dependency that needs to be
injected which is not possible in ConceptsearchParameters.

* [CCP-211] Review comments fix.

* [CCP-211] Changes for review comments.

Made sure correct operator is used while querying lucene.
Number of pages is loaded correctly.

* [CCP-211] Test case fix.

Removed iterating the transient fields using Property descriptor.

* [CCP-218] Changes for displaying merged ids and merged concepts. (#163)

When user searches with the word net id that has been merged, the merged
concept will be displayed, Previously the wrapper of the word net was displayed but from
now the merged concept itself will be displayed.

Also the RESTFul apis will display the merged ids of the concepts.

* [CCP-211] Review comments fixes.

Changed the if conditions to ternary operator.
Fixed a bug in getSimilarTo() in conceptEntry

* [CCP-211] Review comments changes

Added comments on why transient field modifier is used.
Used ternary operator to get the number of records per page.

* [CCP-211] Review comments changes