[CCP-222] Pushed deletion after creating the concept.
As per best practice, deletion is performed after creating the merged
concepts. deletion is performed on the database that existed before
creating the merged concepts. If this is not done, when the code searches
for the word net concept, instead of returning the word net concept, the
system returns the merged concept. The same applies to local CCP concepts
as well.