[CCP-222] Deletion is done before merging (#172)
* [CCP-222] Deletion is done before merging

Deletion is done before merging, because once the concepts are merged, the change made
in the getConceptEntry in ConceptManager always fetches the merged concept. Because of
this instead of fetching the original concept and setting it to deleted, we set the delete
flag on the merged concept.

* [CCP-222] Pushed deletion after creating the concept.

As per best practice, deletion is performed after creating the merged
concepts. deletion is performed on the database that existed before
creating the merged concepts. If this is not done, when the code searches
for the word net concept, instead of returning the word net concept, the
system returns the merged concept. The same applies to local CCP concepts
as well.

* [CCP-222] Passed the concept entry instead of the id

Instead of passing the id and fetching the entry from the id
in delete method, the delete method accepts the concept entry instead
of the id and deletes the corresponding concept entry that has been
passed.

* [CCP-222] Made sure one method is used for deletion of concept