[CCP-227] Changes for deleting word net concepts. (#171)
* [CCP-227] Changes for editing a concept.

When a concept is edited, its wordnet entries are updated
accordingly. Previously when a concept wrapper is edited
the wordnet entries were not updated.

* [CCP-227] Changes for deleting the word net concept

A new method has been introduced in concept manager to delete
the concept directly without fetching the concept and deleting that wrapper.

* [CCP-227] Updates to javadoc review comments

* [CCP-227] Added description for the fields

* [CCP-227] Review comments changes

Instead of passing the concept edit bean, the word net details are passed to the
edit method.

* [CCP-227] Fixed merge conflicts issue

* [CCP-227] Removed the logic of getting entry in client

Client should just retrieve the concept entry. Removed the logic of retrieving
the word net entries and added it to the   manager layers.

* [CCP-227] Review comments changes

* [CCP-227] Review comments fix

* [CCP-227] Review comments fix.

* [CCP-227] Deleted the concept in index instead of updating in client.

* [CCP-227] Renamed method from deleteIndex to deleteFromIndex